+++ src/gui/InvoiceReport.java
@@ -1,18 +1,22 @@
-//Author KAVISHKA
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
 package gui;
 
-import com.raven.datechooser.DateBetween;
-import com.raven.datechooser.DateChooser;
-import com.raven.datechooser.listener.DateChooserAction;
-import com.raven.datechooser.listener.DateChooserAdapter;
-import java.awt.Color;
+import java.awt.BorderLayout;
+import java.awt.Frame;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
 import java.sql.ResultSet;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Vector;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import javax.swing.SwingUtilities;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
 import javax.swing.table.DefaultTableModel;
 import model.MySQL2;
 import net.sf.jasperreports.engine.JRException;
@@ -21,142 +25,198 @@ import net.sf.jasperreports.engine.JasperPrint;
 import net.sf.jasperreports.engine.data.JRTableModelDataSource;
 import net.sf.jasperreports.view.JasperViewer;
 
+/**
+ *
+ * @author Rushma
+ */
 public class InvoiceReport extends javax.swing.JPanel {
     
-    private InvoiceItems invoiceItem;
-    private String id;
     
     private DashboardInterface parent;
     
-    private DateChooser chDate = new DateChooser();
+    private II invoiceItem;
+    private String id;
     
-    private String From;
-    private String To;
-
-    public InvoiceReport(DashboardInterface parent) {
+    private StudentInvoiceHistory invoiceReport;
+    public InvoiceReport() {
         initComponents();
-        dateChooser();
-        loadInvoice("","");
-    }
-    
-    private void dateChooser() {
-        chDate.setTextField(jTextField2);
-        chDate.setDateSelectionMode(DateChooser.DateSelectionMode.BETWEEN_DATE_SELECTED);
-        chDate.setLabelCurrentDayVisible(false);
-        chDate.setForeground(Color.black);
-        chDate.setBackground(Color.white);
-        chDate.setDateFormat(new SimpleDateFormat("yyyy-MM-dd"));
-        chDate.addActionDateChooserListener(new DateChooserAdapter() {
-            @Override
-            public void dateBetweenChanged(DateBetween date, DateChooserAction action) {
-                SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd");
-                From = df.format(date.getFromDate());
-                To = df.format(date.getToDate());
-                loadInvoice(From, To);
-            }
-        });
+        this.invoiceReport = new StudentInvoiceHistory(parent);
+        loadStudentInvoicesHistory(); 
+        addTableMouseListener();
+        addSearchFunctionality();
     }
 
-    private void loadInvoice(String from, String to) {
+    public void loadStudentInvoicesHistory() {
+        
+        String sort = String.valueOf(jComboBox1.getSelectedItem());
+        
+        DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
+        model.setRowCount(0); // Clear the table
 
         try {
-
-            String sort = String.valueOf(jComboBox1.getSelectedItem());
-
-            String searchText = jTextField1.getText().toLowerCase();
-
-            String query = "SELECT * FROM `invoice` "
-                    + "INNER JOIN `student` ON `invoice`.`student_nic` = `student`.`nic` "
-                    + "INNER JOIN `payment_method` ON `invoice`.`payment_method_id` = `payment_method`.`id` "
-                    + "INNER JOIN `invoice_item` ON `invoice`.`id` = `invoice_item`.`invoice_id` "
-                    + "INNER JOIN `courses` ON `invoice_item`.`courses_id` = `courses`.`id` ";
-
-            if (!searchText.isEmpty()) {
-
-//                if (query.contains("WHERE")) {
-//
-//                    query += "LOWER(`class`.`id`) LIKE '%" + searchText + "%' "
-//                            + "OR LOWER(`class`.`name`) LIKE '%" + searchText + "%'";
-//
-//                } else {
-                query += "WHERE (LOWER(`invoice`.`id`) LIKE '%" + searchText + "%' "
-                        + "OR LOWER(`courses`.`name`) LIKE '%" + searchText + "%') ";
-
-//                }
-            }
-
-            if (!from.isEmpty() && to == null) {
-                 query += "AND `invoice`.`date` > '" + from + "' ";
-            } else if (!to.isEmpty() && from == null) {
-                query += "AND `invoice`.`date` < '" + to + "' ";
-            } else if (from != null && !from.isEmpty() && to != null && !to.isEmpty()) {
-                query += "AND `invoice`.`date` > '" + from + "' AND `invoice`.`date` < '" + to + "' ";
-            }
-
+            String query = "SELECT i.id AS invoice_id, "
+                    + "i.student_nic, "
+                    + "CONCAT(s.first_name, ' ', s.last_name) AS student_name, "
+                    + "i.date AS payment_date, "
+                    + "i.total AS amount, "
+                    + "pm.name AS payment_method "
+                    + "FROM invoice i "
+                    + "JOIN student s ON i.student_nic = s.nic "
+                    + "JOIN payment_method pm ON i.payment_method_id = pm.id ";
+            
             if (sort.equals("Invoice ID ASC")) {
-                query += "ORDER BY `invoice`.`id` ASC";
+                query += "ORDER BY `invoice_id` ASC";
             } else if (sort.equals("Invoice ID DESC")) {
-                query += "ORDER BY `invoice`.`id` DESC";
-            } else if (sort.equals("Course Name ASC")) {
-                query += "ORDER BY `courses`.`name` ASC";
-            } else if (sort.equals("Course Name DESC")) {
-                query += "ORDER BY `courses`.`name` DESC";
+                query += "ORDER BY `invoice_id` DESC";
+            } else if (sort.equals("Student NIC ASC")) {
+                query += "ORDER BY `student_nic` ASC";
+            } else if (sort.equals("Student NIC DESC")) {
+                query += "ORDER BY `student_nic` DESC";
             } else if (sort.equals("Cash")) {
-                query += "AND `payment_method`.`name` = 'Cash'";
+                query += "WHERE `pm`.`name` = 'Cash'";
             } else if (sort.equals("Card")) {
-                query += "AND `payment_method`.`name` = 'Card'";
+                query += "WHERE `pm`.`name` = 'Card'";
             }
 
-            ResultSet resultSet = MySQL2.executeSearch(query);
+            System.out.println(query);
+            ResultSet rs = MySQL2.executeSearch(query);
 
-            DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
-            model.setRowCount(0);
+            while (rs.next()) {
+                Vector<String> vector = new Vector <>();
+                vector.add(rs.getString("invoice_id"));
+                vector.add(rs.getString("student_nic"));
+                vector.add(rs.getString("student_name"));
+                vector.add(rs.getString("payment_date"));
+                vector.add(rs.getString("amount"));
+                vector.add(rs.getString("payment_method"));
 
-            while (resultSet.next()) {
-                Vector<String> vector = new Vector<>();
-                vector.add(resultSet.getString("invoice.id"));
-                vector.add(resultSet.getString("courses.name"));
-                vector.add(resultSet.getString("courses.fee"));
-                vector.add(resultSet.getString("student.first_name") + " " + resultSet.getString("student.last_name"));
-                vector.add(resultSet.getString("payment_method.name"));
-                vector.add(resultSet.getString("invoice.date"));
-
+                // Add a row to the table
                 model.addRow(vector);
             }
-
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
+    private void addTableMouseListener() {
+        jTable1.addMouseListener(new MouseAdapter() {
+            public void mouseClicked(MouseEvent e) {
+                if (e.getClickCount() == 2) { // Detect double-click
+                    int selectedRow = jTable1.getSelectedRow();
+                    if (selectedRow != -1) {
+                        String studentNic = jTable1.getValueAt(selectedRow, 1).toString(); // Get Student NIC
+                        openStudentDetailsFrame(studentNic);
+                    }
+                }
+            }
+        });
+    }
+
+    private void openStudentDetailsFrame(String nic) {
+//        int nic = Integer.parseInt(studentNic);
+        StudentDetails detailsFrame = new StudentDetails((Frame) parent, true, nic); // Pass NIC to the frame
+        detailsFrame.setVisible(true);
+    }
+
+    private void addSearchFunctionality() {
+        jTextField1.getDocument().addDocumentListener(new DocumentListener() {
+            @Override
+            public void insertUpdate(DocumentEvent e) {
+                filterTable();
+            }
+
+            @Override
+            public void removeUpdate(DocumentEvent e) {
+                filterTable();
+            }
+
+            @Override
+            public void changedUpdate(DocumentEvent e) {
+                filterTable();
+            }
+        });
+    }
+
+    private void filterTable() {
+        String sort = String.valueOf(jComboBox1.getSelectedItem());
+        String searchText = jTextField1.getText().toLowerCase();
+        DefaultTableModel model = (DefaultTableModel) jTable1.getModel();
+
+        model.setRowCount(0); // Clear the table
+
+        try {
+            String query = "SELECT i.id AS invoice_id, "
+                    + "i.student_nic, "
+                    + "CONCAT(s.first_name, ' ', s.last_name) AS student_name, "
+                    + "i.date AS payment_date, "
+                    + "i.total AS amount, "
+                    + "pm.name AS payment_method "
+                    + "FROM invoice i "
+                    + "JOIN student s ON i.student_nic = s.nic "
+                    + "JOIN payment_method pm ON i.payment_method_id = pm.id "
+                    + "WHERE LOWER(i.student_nic) LIKE '%" + searchText + "%' OR "
+                    + "LOWER(CONCAT(s.first_name, ' ', s.last_name)) LIKE '%" + searchText + "%' ";
+            
+            if (sort.equals("Invoice ID ASC")) {
+                query += "ORDER BY `invoice_id` ASC";
+            } else if (sort.equals("Invoice ID DESC")) {
+                query += "ORDER BY `invoice_id` DESC";
+            } else if (sort.equals("Student NIC ASC")) {
+                query += "ORDER BY `student_nic` ASC";
+            } else if (sort.equals("Student NIC DESC")) {
+                query += "ORDER BY `student_nic` DESC";
+            } else if (sort.equals("Cash")) {
+                query += "AND `pm`.`name` = 'Cash'";
+            } else if (sort.equals("Card")) {
+                query += "AND `pm`.`name` = 'Card'";
+            }
+
+            System.out.println(query);
+            ResultSet rs = MySQL2.executeSearch(query);
+
+            while (rs.next()) {
+                Vector<String> vector = new Vector <>();
+                vector.add(rs.getString("invoice_id"));
+                vector.add(rs.getString("student_nic"));
+                vector.add(rs.getString("student_name"));
+                vector.add(rs.getString("payment_date"));
+                vector.add(rs.getString("amount"));
+                vector.add(rs.getString("payment_method"));
+
+                // Add a row to the table
+                model.addRow(vector);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
-        jLabel1 = new javax.swing.JLabel();
         jScrollPane1 = new javax.swing.JScrollPane();
         jTable1 = new javax.swing.JTable();
-        jLabel2 = new javax.swing.JLabel();
-        jComboBox1 = new javax.swing.JComboBox<>();
-        jLabel5 = new javax.swing.JLabel();
+        jLabel1 = new javax.swing.JLabel();
         jTextField1 = new javax.swing.JTextField();
-        jLabel8 = new javax.swing.JLabel();
-        jButton2 = new javax.swing.JButton();
-        jTextField2 = new javax.swing.JTextField();
+        jLabel2 = new javax.swing.JLabel();
+        jLabel3 = new javax.swing.JLabel();
+        jButton1 = new javax.swing.JButton();
         jButton3 = new javax.swing.JButton();
-
-        jLabel1.setFont(new java.awt.Font("Century Gothic", 1, 26)); // NOI18N
-        jLabel1.setText("Invoice Report");
+        jLabel4 = new javax.swing.JLabel();
+        jComboBox1 = new javax.swing.JComboBox<>();
 
         jTable1.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null},
-                {null, null, null, null, null, null}
+
             },
             new String [] {
-                "Invoice ID", "Course Name", "Amount Paid", "Student Name", "Payment Method", "Payment Date"
+                "Invoice ID", "Student NIC", "Student Name", "Payment date", "Amount", "Payment Method"
             }
         ) {
             boolean[] canEdit = new boolean [] {
@@ -174,33 +234,25 @@ public class InvoiceReport extends javax.swing.JPanel {
         });
         jScrollPane1.setViewportView(jTable1);
 
+        jLabel1.setFont(new java.awt.Font("Century Gothic", 1, 24)); // NOI18N
+        jLabel1.setText("Invoice Report");
+
+        jTextField1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jTextField1ActionPerformed(evt);
+            }
+        });
+
         jLabel2.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel2.setText("Sort By :");
+        jLabel2.setText("Search by Name or NIC");
 
-        jComboBox1.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Invoice ID ASC", "Invoice ID DESC", "Course Name ASC", "Course Name DESC", "Cash", "Card" }));
-        jComboBox1.addActionListener(new java.awt.event.ActionListener() {
+        jLabel3.setText("Double click a row to view student details");
+
+        jButton1.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
+        jButton1.setText("Print");
+        jButton1.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jComboBox1ActionPerformed(evt);
-            }
-        });
-
-        jLabel5.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel5.setText("Search");
-
-        jTextField1.addKeyListener(new java.awt.event.KeyAdapter() {
-            public void keyReleased(java.awt.event.KeyEvent evt) {
-                jTextField1KeyReleased(evt);
-            }
-        });
-
-        jLabel8.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jLabel8.setText("Sort By Date");
-
-        jButton2.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
-        jButton2.setText("Print");
-        jButton2.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton2ActionPerformed(evt);
+                jButton1ActionPerformed(evt);
             }
         });
 
@@ -212,83 +264,77 @@ public class InvoiceReport extends javax.swing.JPanel {
             }
         });
 
+        jLabel4.setFont(new java.awt.Font("Century Gothic", 1, 14)); // NOI18N
+        jLabel4.setText("Sort By :");
+
+        jComboBox1.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Invoice ID ASC", "Invoice ID DESC", "Student NIC ASC", "Student NIC DESC", "Cash", "Card" }));
+        jComboBox1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jComboBox1ActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jLabel1)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
             .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
+                .addGap(16, 16, 16)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGap(0, 0, Short.MAX_VALUE)
-                        .addComponent(jLabel1)
-                        .addGap(0, 0, Short.MAX_VALUE))
-                    .addComponent(jScrollPane1)
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(jLabel2)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 206, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(18, 18, 18)
+                        .addComponent(jLabel4)
                         .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                         .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 29, Short.MAX_VALUE)
-                        .addComponent(jLabel5)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 162, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 29, Short.MAX_VALUE)
-                        .addComponent(jLabel8)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 161, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                         .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, 123, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(10, 10, 10)
-                        .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap())
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jButton1))
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 968, Short.MAX_VALUE)
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(jLabel3)
+                        .addGap(0, 0, Short.MAX_VALUE)))
+                .addGap(16, 16, 16))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(19, 19, 19)
+                .addContainerGap(24, Short.MAX_VALUE)
                 .addComponent(jLabel1)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 20, Short.MAX_VALUE)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel2)
-                    .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 21, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel8, javax.swing.GroupLayout.PREFERRED_SIZE, 38, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jTextField2, javax.swing.GroupLayout.PREFERRED_SIZE, 30, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jButton2, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(24, 24, 24)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 443, Short.MAX_VALUE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 52, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(jLabel4)
+                        .addComponent(jComboBox1, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 21, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGap(19, 19, 19)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 389, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jLabel3)
+                .addContainerGap())
         );
+
+        jTextField1.getAccessibleContext().setAccessibleDescription("");
     }// </editor-fold>//GEN-END:initComponents
 
-    private void jComboBox1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox1ActionPerformed
-        if (From == null && To != null) {
-            loadInvoice(From,"");
-        } else if (From != null && To == null) {
-            loadInvoice("",To);
-        } else if (From != null && To != null) {
-            loadInvoice(From,To);
-        } else {
-            loadInvoice("","");
-        }
-    }//GEN-LAST:event_jComboBox1ActionPerformed
+    private void jTextField1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jTextField1ActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jTextField1ActionPerformed
 
-    private void jTextField1KeyReleased(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_jTextField1KeyReleased
-        if (From == null && To != null) {
-            loadInvoice(From,"");
-        } else if (From != null && To == null) {
-            loadInvoice("",To);
-        } else if (From != null && To != null) {
-            loadInvoice(From,To);
-        } else {
-            loadInvoice("","");
-        }
-    }//GEN-LAST:event_jTextField1KeyReleased
-
-    private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
+    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
         //view or print report
-        String path = "src//reports//StudentInvoiceReport.jasper";
+        String path = "src//reports//StudentInvoiceHistory.jasper";
         
         String dateTime = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(new Date());
         
@@ -305,31 +351,49 @@ public class InvoiceReport extends javax.swing.JPanel {
         }
 
         JasperViewer.viewReport(jasperPrint, false);
-    }//GEN-LAST:event_jButton2ActionPerformed
+    }//GEN-LAST:event_jButton1ActionPerformed
 
     private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton3ActionPerformed
-        InvoiceItems ii = new InvoiceItems(id);
-        ii.setVisible(true);
+        InvoiceItems invoiceItems = new InvoiceItems(parent, true, id);
+        invoiceItems.setVisible(true);
     }//GEN-LAST:event_jButton3ActionPerformed
 
     private void jTable1MouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTable1MouseClicked
-        int selectedRow = jTable1.getSelectedRow();
+        if (evt.getClickCount() == 1) {
+  
+            int selectedRow = jTable1.getSelectedRow();
             
             id = String.valueOf(jTable1.getValueAt(selectedRow, 0));
+//            invoiceItem.setId(id);
+            
+        }
     }//GEN-LAST:event_jTable1MouseClicked
+
+    private void jComboBox1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox1ActionPerformed
+//        if (From == null && To != null) {
+//            loadInvoice(From,"");
+//        } else if (From != null && To == null) {
+//            loadInvoice("",To);
+//        } else if (From != null && To != null) {
+//            loadInvoice(From,To);
+//        } else {
+//            loadInvoice("","");
+//        }
+        loadStudentInvoicesHistory();
+        filterTable();
+    }//GEN-LAST:event_jComboBox1ActionPerformed
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton jButton2;
+    private javax.swing.JButton jButton1;
     private javax.swing.JButton jButton3;
     private javax.swing.JComboBox<String> jComboBox1;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel8;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JTable jTable1;
     private javax.swing.JTextField jTextField1;
-    private javax.swing.JTextField jTextField2;
     // End of variables declaration//GEN-END:variables
 }
